requirements.txt
# Core - using versions with pre-built Windows wheels
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
yfinance>=0.2.30
plotly>=5.17.0

# Options pricing
mibian==0.1.3

# Alpaca API
alpaca-py>=0.20.0

# LLM
openai>=1.10.0
anthropic>=0.18.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
